Some FIND, an example not required, implemented the tree and minor things
FIND: how to convert hash to array and backwards

Implemented the 16 elements print with "inject". The problem asked with "each",
wich is already done. But with "inject" i could print the "residual" for
collections with a number of elements not multiple of 4.

Implemented the Tree problem with a recursive "initialize".

*grep*

Used the block version of open, is just like the "with" of Python.
Used the each_line iterator, instead of reading the whole file with readlines.